feat: Fix Background Music Bugs
- Created a SwabberMusic Class to contain all music-related parameters
and functions
- Fixed mute/unmute functionality: pressing the button
mutes the music (but it still runs in the background), while leaving the
app pauses it.
- Added fade in/out functions in the SwabberMusic Class
- Restarting the music from the beginning if returning to the home screen
from the game
<BUG>: if the music was muted in the home screen, and then
in the game you return to home screen from the pause screen - the music
is unmuted. This is because when returning to the home screen from the
pause screen, *THE ACTIVITY IS CREATED AGAIN!* (onCreate is called)