Custom Mode Version 2: Value Validation
- added a validation for the values inserted in the text boxes, to check
that they are in the range of [MIN, MAX], specified for each field
individualy
- if the inserted value is not in the range, the text is
painted red and a toast appears, describing the MIN MAX values for the
current field
- The game cannot be started until all the fields are
white
- some of the MIN MAX values are dependent on other fields and
change dynamically when the values are changed
- The validatiion is
realazied in the TextValidator class, which is added as a listener to
each text box

## KNOWN BUG: if a value is set out of range, and is
painted red, and then other fields change such that now it is inside a
valid range (because of the dynamic ranges), the value will stay red and
will not be validated until it is inserted again.
This is because when a
value is changed, only itself is validated, and not all other fields. ##